requirements.txt
aiohttp==3.11.16
aiolimiter==1.2.1
python-dotenv==1.1.0
